Remove :orphan: directives
There are no more orphans, and we don't want them.
Any new page which does not appear in the site's TOC
will display an error when building the docs.

This is committed separatedly because it touches 80+
files, but it does nothing more than remove the top two lines
of each.

The contributing guideline has been updated so that it does
not suggest that :orphan: is added to new pages.