feat: Remove empty TOCs from sidebar
It just displayed "On this page", with no content.

The issue is that the localtoc always includes the page title, which is redundant,
so it is removed via CSS.
This makes detecting and removing empty TOCs a bit complex
(because there might be a hidden title).
On top of that, what reaches the layout is rendered HTML, so detection is
based on that. Ugly.