feat: Do not give inline structs their own headers
To be precise: structs which are used inlined inside some other struct are not
usually needed on their own, so do not give them their own header.
They still appear inside all containing structs (with the grayed out title).

This reduces a lot the number of headers, but still looks too long...