fix: Example code language selector tabs
Something broken when updated Sphinx.
There was space between the tabs and the example code, and the word "notranslate"
was present in the tab text AND its class, so it didn't work.
I don't know how to properly fix this (and I suspect the change comes from
the viewsource plugin) so I provide this horrible hack, where everything is
patched up at runtime ¬.¬'