WIP: Add glossary tooltips
The idea is to manually add :term: to all words that have a glossary definition.
Sphinx already handles this nicely. What I want to do is
add a tooltip with the definition so you don't need to move to a
different page. There's a beautiful extension that does all this,
sphinx-hoverxref, but only works with docs hosted on readthedocs.com