fix: Remove remaining references to NGL nodes
We've shut down all NGL nodes (mainnet and testnet)
so they cannot be referenced from the docs.
But we cannot reference some random community node either.
The proper mechanism is to use the Statistics Service, which is hinted
by this commit. Later on we should develop a whole page for it
(just like the Faucet and the Explorer).